gameserver/requirements.txt
python-socketio
eventlet
openai
azure-data-tables
